requirements.txt
pymed
gradio
haystack-ai
python-dotenv
transformers